+++ NWaves/FeatureExtractors/Multi/TimeDomainFeaturesExtractor.cs
@@ -24,14 +24,14 @@ namespace NWaves.FeatureExtractors.Multi
         public override int FeatureCount => FeatureDescriptions.Length;
 
         /// <summary>
-        /// Length of analysis window (in ms)
+        /// Length of analysis window (in seconds)
         /// </summary>
-        private readonly float _windowSize;
+        private readonly double _windowSize;
 
         /// <summary>
         /// Hop length (in ms)
         /// </summary>
-        private readonly float _hopSize;
+        private readonly double _hopSize;
 
         /// <summary>
         /// Extractor functions
@@ -46,7 +46,7 @@ namespace NWaves.FeatureExtractors.Multi
         /// <param name="windowSize"></param>
         /// <param name="hopSize"></param>
         public TimeDomainFeaturesExtractor(string featureList,
-                                           float windowSize = 0.0256f, float hopSize = 0.010f,
+                                           double windowSize = 0.0256/*sec*/, double hopSize = 0.010/*sec*/,
                                            IReadOnlyDictionary<string, object> parameters = null)
         {
             if (featureList == "all" || featureList == "full")
@@ -115,7 +115,7 @@ namespace NWaves.FeatureExtractors.Multi
                 featureVectors.Add(new FeatureVector
                 {
                     Features = featureVector,
-                    TimePosition = (float)i / signal.SamplingRate
+                    TimePosition = (double)i / signal.SamplingRate
                 });
 
                 i += hopSize;
