+++ NWaves/FeatureExtractors/SpnccExtractor.cs
@@ -64,14 +64,14 @@ namespace NWaves.FeatureExtractors
         private readonly int _fftSize;
 
         /// <summary>
-        /// Length of analysis window (in ms)
+        /// Length of analysis window (in seconds)
         /// </summary>
-        private readonly float _windowSize;
+        private readonly double _windowSize;
 
         /// <summary>
-        /// Hop length (in ms)
+        /// Hop length (in seconds)
         /// </summary>
-        private readonly float _hopSize;
+        private readonly double _hopSize;
 
         /// <summary>
         /// Type of the window function
@@ -99,7 +99,7 @@ namespace NWaves.FeatureExtractors
         /// <param name="window"></param>
         public SpnccExtractor(int featureCount, int power = 15,
                              int filterbankSize = 40, float lowFreq = 100, float highFreq = 6800,
-                             float windowSize = 0.0256f, float hopSize = 0.010f, int fftSize = 1024,
+                             double windowSize = 0.0256/*sec*/, double hopSize = 0.010/*sec*/, int fftSize = 1024,
                              float preEmphasis = 0.0f, WindowTypes window = WindowTypes.Hamming)
         {
             FeatureCount = featureCount;
@@ -261,7 +261,7 @@ namespace NWaves.FeatureExtractors
                 featureVectors.Add(new FeatureVector
                 {
                     Features = spnccs,
-                    TimePosition = (float)i / signal.SamplingRate
+                    TimePosition = (double)i / signal.SamplingRate
                 });
 
                 i += hopSize;
